feat: support path mapping for conda install rules
The basic idea with path mapping is that rather than execute an action
with input/output paths that look like

    bazel-out/k8-fastbuild-abc123/bin

which combines the build mode, architecture, and a hash of the
configuration state, have the action executing in the sandbox (or remote
execution) use a path like

    bazel-out/cfg/bin

Very frequently (including in the conda installer rule) that path is the
only difference between the actions executing for different
configurations, so doing that mapping means that it can share the action
cache key between several target configurations, improving cache hit
rates and therefore load on the remote execution service.

See https://github.com/bazelbuild/bazel/discussions/22658 for further
details, both on why it's useful and why it required these changes.